% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/StudyInvariant.R
\name{email}
\alias{email}
\title{Email results}
\usage{
email(from, to = getDestinationAddress(),
  subject = getDefaultStudyEmailSubject(), dataDescription,
  file = getDefaultStudyFileName())
}
\arguments{
\item{from}{Return email address}

\item{to}{Delivery email address (must be a gmail.com acccount)}

\item{subject}{Subject line of email}

\item{dataDescription}{A short description of the database}

\item{file}{(Optional) Name of local file with results; makee sure to use forward slashes (/)}
}
\value{
A list of files that were emailed.
}
\description{
Email results
}
\details{
This function emails the result CSV files to the study coordinator.
}

