% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/StudySpecific.R
\name{execute}
\alias{execute}
\title{Execute OHDSI Study StableDM2Patients}
\usage{
execute(dbms, user, password, server, port = NULL, cdmSchema, resultsSchema,
  file = getDefaultStudyFileName(), ...)
}
\arguments{
\item{dbms}{The type of DBMS running on the server. Valid values are
\itemize{
  \item{"mysql" for MySQL}
  \item{"oracle" for Oracle}
  \item{"postgresql" for PostgreSQL}
  \item{"redshift" for Amazon Redshift}
  \item{"sql server" for Microsoft SQL Server}
}}

\item{user}{The user name used to access the server.}

\item{password}{The password for that user}

\item{server}{The name of the server}

\item{port}{(optional) The port on the server to connect to}

\item{cdmSchema}{Schema name where your patient-level data in OMOP CDM format resides}

\item{resultsSchema}{(Optional) Schema where you'd like the results tables to be created (requires user to have create/write access)}

\item{file}{(Optional) Name of local file to place results; makre sure to use forward slashes (/)}
}
\value{
Study results are placed in CSV format files in specified local folder and returned
as an R object class \code{OhdsiStudy} when sufficiently small.  The properties of an
\code{OhdsiStudy} may differ from study to study.
}
\description{
Execute OHDSI Study StableDM2Patients
}
\details{
This function executes OHDSI Study Estimating the number of stable diabetes mellitus type II patients.
This is a study of Guidelines for several chronic conditions, like hypertension or depression, still encompass a large amount of uncertainty, often leaving the physician with several choices and not enough tools to decide between them.
Our aim is to introduce a data-driven approach that suggests a treatment based on similarity of a new patient to patients that received a beneficial treatments in the past. In order to assess the utility of our approach, we test it on diabetes mellitus type II (DM) which has relatively good guidelines.
Assessing stability of DM patients requires longitudinal information including also outpatient data. As a first step, we estimate the number of patients that are stable on known treatments for diabetes.
The query assesses the number of patients that are considered stable on any treatment, by posing several requirements on the amount of information we have on the patient, including total time of observational data, number of visits and the number of relevant lab tests as well as verifying that the patient is on a certain drug above a minimal amount of time.
Before coding the complete study, we are releasing the potential patient identification step to determine if there is enough support on the OHDSI sites for this study to be ported in its entirety.
}
\examples{
\dontrun{
# Run study
execute(dbms = "postgresql",
        user = "joebruin",
        password = "supersecret",
        server = "myserver",
        cdmSchema = "cdm_schema",
        resultsSchema = "results_schema")

# Email result file
email(from = "collaborator@ohdsi.org",
      dataDescription = "StableDM2Patients results")
}
}

